refactor(iopoll): split internal poll implementation
