+++ src/Reseau.jl
@@ -16,7 +16,7 @@ include("0_memory_compat.jl")
 include("0_gcsafe_ccall_compat.jl")
 include("1_eventloops.jl")
 include("2_socket_ops.jl")
-include("3_internal_poll.jl")
+include("3_iopoll.jl")
 include("4_tcp.jl")
 include("5_host_resolvers.jl")
 include("6_tls.jl")
